Minimum Number of Operations to Move All Balls
Firstly we find the number of operations to at 0th index and we store number of one's after the particular index in ahead variable and number of one's behind the index we store in behind variable then we iterate loop for every index by res[i] = res[i-1]-ahead +behind logical formula. we then written the final answer.