1578. Minimum Time to Make Rope Colorful
Why we updated cost with max value ??
Ans =>Element with minimum value is deleted which means element having maximum value is kept. For example if we have 4,3,5 them we should delete 4 and 3 so the count will be 7 if we don't update neededTime[i] i.e., cost value to max then it will be 3 + 3 = 6 we get wrong answer. Thus, adding max value to the cost.